test(e2e): unify upload tests and add npm test + CI workflow
Extract shared boilerplate (~70 lines per file) from the three upload
test scripts into test-helpers.mjs. Unify default PORT to 8080 across
all test scripts. Add npm scripts (test, test:e2e, test:upload, etc.)
for convenient local execution. Replace model-registry-test.yml with
a unified e2e.yml workflow that runs all E2E tests with weekly schedule
and failure artifact upload.
