--- .github/workflows/model-registry-test.yml
@@ -1,40 +1,0 @@
-name: Model Registry Smoke Test
-
-on:
-  pull_request:
-  push:
-    branches:
-      - main
-
-jobs:
-  smoke-test:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-
-      - uses: actions/setup-node@v4
-        with:
-          node-version: "20"
-          cache: "npm"
-
-      - name: Install dependencies
-        run: |
-          npm ci
-          npx playwright install chromium --with-deps
-
-      - name: Start local server
-        run: |
-          python3 -m http.server 3000 &
-          # Wait for server to be ready (up to 30 seconds)
-          for i in {1..30}; do
-            if curl -s http://localhost:3000/index.html > /dev/null 2>&1; then
-              echo "Server ready"
-              break
-            fi
-            echo "Waiting for server... ($i)"
-            sleep 1
-          done
-          curl -sf http://localhost:3000/index.html > /dev/null || exit 1
-
-      - name: Run Model Registry smoke tests
-        run: node scripts/model-registry-smoke.mjs
